refactor/buildSrc module name correction (#558)
* Correct buildSrc module module name type

* Update version.json

* Remove dangling import
