+++ buildSrc/src/main/java/com/mxt/anitrend/buildsrc/components/PluginComponents.kt
@@ -1,4 +1,4 @@
-package com.mxt.anitrend.buildsrc.componets
+package com.mxt.anitrend.buildsrc.components
 
 import com.mxt.anitrend.buildsrc.extensions.baseExtension
 import org.gradle.api.Project
