+++ buildSrc/src/main/java/com/mxt/anitrend/buildsrc/extensions/ProjectExtensions.kt
@@ -19,7 +19,7 @@ package com.mxt.anitrend.buildsrc.extensions
 
 import com.android.build.gradle.BaseExtension
 import com.android.build.gradle.internal.dsl.BaseAppModuleExtension
-import com.mxt.anitrend.buildsrc.componets.PropertiesReader
+import com.mxt.anitrend.buildsrc.components.PropertiesReader
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.kotlin.dsl.getByType
