+++ buildSrc/src/main/java/com/mxt/anitrend/buildsrc/plugin/CorePlugin.kt
@@ -17,9 +17,12 @@
 
 package com.mxt.anitrend.buildsrc.plugin
 
+import com.mxt.anitrend.buildsrc.components.applyAndroidConfiguration
+import com.mxt.anitrend.buildsrc.components.configureAdditionalPlugins
+import com.mxt.anitrend.buildsrc.components.configureDependencies
+import com.mxt.anitrend.buildsrc.components.configurePlugins
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import com.mxt.anitrend.buildsrc.componets.*
 
 open class CorePlugin : Plugin<Project> {
